refactor: added all player stats to the database

+++ src/main/java/me/tech/foraging/models/database/DBStats.java
@@ -18,8 +18,8 @@ public class DBStats {
 		DBStats dbStats = new DBStats();
 
 		dbStats.uuid = result.getString("uuid");
-		dbStats.health = result.getDouble("health");
-		dbStats.damage = result.getDouble("damage");
+		dbStats.health = result.getInt("health");
+		dbStats.damage = result.getInt("damage");
 
 		return dbStats;
 	}
